Remove serverside StandingStateSystem (#35999)
Refactor system
